+++ src/game/behaviors/flame.inc.c
@@ -22,8 +22,10 @@ void bhv_small_piranha_flame_loop(void) {
 
         obj_compute_vel_from_move_pitch(o->oSmallPiranhaFlameStartSpeed);
         cur_obj_move_standard(-78);
-        spawn_object_with_scale(o, o->oSmallPiranhaFlameModel, bhvSmallPiranhaFlame,
-                                0.4f * o->header.gfx.scale[0]);
+        if (!configVrFlameOptimizations || (o->oTimer & 1) == 0) {
+            spawn_object_with_scale(o, o->oSmallPiranhaFlameModel, bhvSmallPiranhaFlame,
+                                    (configVrFlameOptimizations ? 0.55f : 0.4f) * o->header.gfx.scale[0]);
+        }
 
         if (o->oTimer > o->oSmallPiranhaFlameNextFlameTimer) {
             spawn_object_relative_with_scale(1, 0, o->oGraphYOffset, 0, o->header.gfx.scale[0], o,