+++ src/pc/djui/djui_panel_vr.c
@@ -89,6 +89,7 @@ static void djui_panel_vr_performance_defaults(struct DjuiBase* caller) {
 
     configVrRenderScale = 100;
     configVrShowFps = false;
+    configVrFlameOptimizations = false;
     configVrDesktopMirror = true;
     configVrDesktopMirrorFps = 60;
 }
@@ -388,6 +389,13 @@ static void djui_panel_vr_performance_create(struct DjuiBase* caller) {
             body,
             "FPS Counter",
             &configVrShowFps,
+            NULL
+        );
+
+        djui_checkbox_create(
+            body,
+            "Flame & Lava Optimizations",
+            &configVrFlameOptimizations,
             NULL
         );